3.34 Update
water mobs are no longer going to spawn (except for the few required), ARMY PIGMAN will be nerfed to only a few spawn and not an "army" of them, added the long requested mod Thaumic Inventory Scanning, and lastly Draconic Reactor (yes this bad boi) will be getting a massive buff. Currently it can only produce in the 2 millions which is WEAK. I have 50x the power rate so it can be useful again. (PLEASE give me feedback on the power rate [is it to much or not enough for the work required?])